requirements.txt
ebrains-kg-core
